Starting work on switching to classes workflow
