+++ setup.cfg
@@ -16,7 +16,14 @@ universal = 1
 
 [flake8]
 exclude = docs
-max-line-length = 120
+max-line-length = 88
+ignore =
+    # Whitespace before ':'
+    E203
+    # Whitepsace at end of line
+    W291
+    # Line break before logical
+    W503
 
 [aliases]
 # Define setup.py command aliases here
