+++ setup.py
@@ -49,6 +49,7 @@ requirements = [
     "aicspylibczi==2.1.0",
     "dask==2.8.1",
     "dask_jobqueue==0.7.0",
+    "fsspec==0.6.1",
     "imageio==2.6.1",
     "imageio-ffmpeg==0.3.0",
     "lxml==4.4.1",
@@ -70,8 +71,8 @@ extra_requirements = {
         *test_requirements,
         *setup_requirements,
         *dev_requirements,
-        *interactive_requirements
-    ]
+        *interactive_requirements,
+    ],
 }
 
 setup(
