--- timelapse_tools/normalization/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Normalization package for timelapse_tools."""
