+++ tox.ini
@@ -1,5 +1,7 @@
 [flake8]
-exclude = timelapse_tools/vendor/*
+exclude =
+    timelapse_tools/vendor/*
+    timelapse_tools/outdated/*
 
 [tox]
 skipsdist = True
